# Conflicts:
#	miniproject/src/main/java/com/sparta/miniproject/dto/PostRequestDto.java

+++ miniproject/src/main/java/com/sparta/miniproject/service/CommentService.java
@@ -16,15 +16,21 @@ import java.util.List;
 public class CommentService {
 
     private final CommentRepository commentRepository;
+    private final UserService userService;
 
     // 댓글 조회
     public List<Comment> getComment(Long postId) {
         return commentRepository.findAllByOrderByCreatedAtDesc(postId);
     }
 
+    private String getUser() {
+        return userService.getMyInfo().getUsername();
+    }
+
     // 댓글 생성
     @Transactional
     public Comment createComment(Long postId, CommentRequestDto requestDto) {
+        requestDto.setAuthor(getUser());
         return commentRepository.save(new Comment(postId, requestDto));
     }
 
@@ -33,6 +39,9 @@ public class CommentService {
     public Comment updateComment(Long postId, Long commentId, CommentRequestDto requestDto) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 댓글입니다."));
+        if(!getUser().equals(comment.getAuthor())) {
+            throw new IllegalArgumentException("작성자가 아닙니다.");
+        }
         comment.update(requestDto);
         return comment;
     }
@@ -41,6 +50,9 @@ public class CommentService {
     public void deleteComment(Long commentId) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 댓글입니다"));
+        if(!getUser().equals(comment.getAuthor())) {
+            throw new IllegalArgumentException("작성자가 아닙니다.");
+        }
         commentRepository.deleteById(commentId);
     }
 }
